refactor: abstract call for calling methods
