+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export type { InjectedCkb, CkbProvider } from './injected';
-export type { OwnershipService, NotificationService, KeystoreService, GrantService } from './services';
+export type { OwnershipService, NotificationService, KeystoreService, GrantService, ConfigService } from './services';
 export type { Storage } from './storage';
 export type { Promisable, Paginate, Cursor } from './base';
+export type { Call, CallMap, CallResult, CallParam } from './call';
